Organized the project into multiple files.  - It adds more readibility  - Makes the codebase more convenient to work on  - Later on it will be even more useful
